move helper functions from rewards v1 test to rewards v2 test
